Ensuring absolute path for assets
On some servers (Netlify in my case), styles will not load and return an error complaining:
`...[filename]is not a supported stylesheet MIME type, and strict MIME checking is enabled.`

The [fix](https://stackoverflow.com/a/34628034) is apparently setting a baseURL in the meta tag (this didn't work for me) or setting the `publicPath: '/'` to server root. The latter seemed to work for me.